feat(structure): Remove src folder and move funciton under lib
