+++ jest.config.js
@@ -68,7 +68,7 @@ module.exports = {
 	// An array of directory names to be searched recursively up from the requiring module's location
 	moduleDirectories: [
 		'node_modules',
-		'src'
+		'lib'
 	],
 
 	// An array of file extensions your modules use
