+++ lib/input.js
@@ -1,9 +1,9 @@
-/* src/index.js */
+/* lib/input.js */
 'use strict';
 
 const fs = require('fs');
 
-const { inputFromSTDIN } = require('./util');
+const { inputFromSTDIN } = require('./utils');
 const splitter = require('../lib/splitter');
 const matcher = require('../lib/matcher');
 const Analyzer = require('../lib/analyzer');
