+++ package.json
@@ -10,8 +10,8 @@
 	"scripts": {
 		"start": "NODE_ENV=development node ./pgfilter.js",
 		"debug": "NODE_PATH=. node --inspect-brk=0.0.0.0:9229 ./pgfilter.js",
-		"test": "NODE_ENV=testing NODE_PATH=src jest test",
-		"coverage": "NODE_ENV=testing NODE_PATH=src jest test --coverage",
+		"test": "NODE_ENV=testing NODE_PATH=lib jest test",
+		"coverage": "NODE_ENV=testing NODE_PATH=lib jest test --coverage",
 		"lint": "eslint ."
 	},
 	"repository": {
@@ -28,7 +28,6 @@
 	"files": [
 		"pgfilter.js",
 		"lib/",
-		"src/",
 		"README.md",
 		"LICENSE"
 	],
